+++ app/Config/Constants.php
@@ -97,3 +97,4 @@ define('HTTP_STATUS_OK', 200);
 define('HTTP_STATUS_BAD_REQUEST', 400);
 define('HTTP_STATUS_UNAUTHORIZED', 401);
 define('HTTP_STATUS_NOT_FOUND', 404);
+define('HTTP_INTERNAL_SERVER_ERROR', 500);
