+++ app/Config/Filters.php
@@ -2,6 +2,7 @@
 
 namespace Config;
 
+use App\Filters\AuthToken;
 use CodeIgniter\Config\BaseConfig;
 use CodeIgniter\Filters\CSRF;
 use CodeIgniter\Filters\DebugToolbar;
@@ -23,6 +24,7 @@ class Filters extends BaseConfig
         'honeypot'      => Honeypot::class,
         'invalidchars'  => InvalidChars::class,
         'secureheaders' => SecureHeaders::class,
+        'authlogin'     => AuthToken::class
     ];
 
     /**
